Issue #43, #53: Fix libzip compile issues
Modified the define check to look for LIBZIP_VERSION as opposed to ZIP_CODEC_ENCODE.  libzip 0.10 introduced LIBZIP_VERSION (which is also used by all later versions) and also introduced the defines that the guard handles.  ZIP_CODEC_ENCODE is no longer used by later libzip versions (e.g. 1.12) which was causing compilation errors.

Also, the config.ac has a check that libzip is >= 0.10 so this is mostly a moot point.